Create Parser.java to parse user commands
