+++ src/main/java/InvalidTaskException.java
@@ -1,5 +1,5 @@
 public class InvalidTaskException extends Exception {
-    public InvalidTaskException(Gray.TaskType taskType, Gray.MissingInfo missingInfo) {
+    public InvalidTaskException(Parser.TaskType taskType, Parser.MissingInfo missingInfo) {
         super("Please provide a " + missingInfo.getMissingInfo() + " for your " + taskType.getTaskType() + ".");
     }
 }
